Move Attribute to its own file, too.
